Rename StatusService to StatusApiService
