+++ src/main/scala/org/broadinstitute/dsde/firecloud/webservice/StatusApiService.scala
@@ -1,22 +1,20 @@
-package org.broadinstitute.dsde.firecloud.service
+package org.broadinstitute.dsde.firecloud.webservice
 
 import java.text.SimpleDateFormat
 import java.util.Date
 
 import akka.actor.Actor
+import org.broadinstitute.dsde.firecloud.FireCloudConfig
+import org.broadinstitute.dsde.firecloud.service.{FireCloudDirectives, MethodsService, PerRequestCreator}
 import spray.json._
-
 import spray.routing._
 
-
-import org.broadinstitute.dsde.firecloud.FireCloudConfig
-
-class StatusServiceActor extends Actor with StatusService {
+class StatusApiServiceActor extends Actor with StatusApiService {
   def actorRefFactory = context
   def receive = runRoute(routes)
 }
 
-trait StatusService extends HttpService with PerRequestCreator with FireCloudDirectives {
+trait StatusApiService extends HttpService with PerRequestCreator with FireCloudDirectives {
 
   private final val ApiPrefix = "status"
   private final val dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
